refactor, code cleanup
