+++ cs3api4lab/handlers.py
@@ -11,7 +11,7 @@ from cs3api4lab.api.cs3_public_share_api import Cs3PublicShareApi
 from cs3api4lab.api.cs3_user_api import Cs3UserApi
 from cs3api4lab.api.cs3_file_api import Cs3FileApi
 from jupyter_server.utils import url_path_join
-
+from cs3api4lab.utils.asyncify import get_or_create_eventloop
 
 class ShareHandler(APIHandler):
     @property
@@ -23,34 +23,34 @@ class ShareHandler(APIHandler):
     def post(self):
         request = self.get_json_body()
         try:
-            yield AsyncRequestHandler.handle_request(self,
-                                          self.share_api.create,
-                                          201,
-                                          request['endpoint'],
-                                          request['file_path'],
-                                          request['grantee'],
-                                          request['idp'],
-                                          request['role'],
-                                          request['grantee_type'])
+            yield RequestHandler.async_handle_request(self,
+                                                      self.share_api.create,
+                                                      201,
+                                                      request['endpoint'],
+                                                      request['file_path'],
+                                                      request['grantee'],
+                                                      request['idp'],
+                                                      request['role'],
+                                                      request['grantee_type'])
         except KeyError as err:
-            AsyncRequestHandler.handle_error(self, ParamError(err))
+            RequestHandler.handle_error(self, ParamError(err))
 
     @web.authenticated
     @gen.coroutine
     def delete(self):
         share_id = self.get_query_argument('share_id')
-        yield AsyncRequestHandler.handle_request(self, self.share_api.remove, 204, share_id)
+        yield RequestHandler.async_handle_request(self, self.share_api.remove, 204, share_id)
 
     def put(self):
         request = self.get_json_body()
         try:
-            yield AsyncRequestHandler.handle_request(self,
-                                          self.share_api.update_share,
-                                          204,
-                                          request['share_id'],
-                                          request['role'])
+            yield RequestHandler.async_handle_request(self,
+                                                      self.share_api.update_share,
+                                                      204,
+                                                      request['share_id'],
+                                                      request['role'])
         except KeyError as err:
-            AsyncRequestHandler.handle_error(self, ParamError(err))
+            RequestHandler.handle_error(self, ParamError(err))
 
 
 class ListSharesHandler(APIHandler):
@@ -61,7 +61,7 @@ class ListSharesHandler(APIHandler):
     @web.authenticated
     @gen.coroutine
     def get(self):
-        yield AsyncRequestHandler.handle_request(self, self.share_api.list_shares, 200)
+        yield RequestHandler.async_handle_request(self, self.share_api.list_shares, 200)
 
 
 class ListReceivedSharesHandler(APIHandler):
@@ -74,13 +74,13 @@ class ListReceivedSharesHandler(APIHandler):
     def get(self):
         status = self.get_query_argument('status', default=None)
 
-        yield AsyncRequestHandler.handle_request(self, self.share_api.list_received, 200, status)
+        yield RequestHandler.async_handle_request(self, self.share_api.list_received, 200, status)
 
     @web.authenticated
     @gen.coroutine
     def put(self):
         body = self.get_json_body()
-        yield AsyncRequestHandler.handle_request(self, self.share_api.update_received, 200, body["share_id"], body["state"])
+        yield RequestHandler.async_handle_request(self, self.share_api.update_received, 200, body["share_id"], body["state"])
 
 
 class ListSharesForFile(APIHandler):
@@ -92,7 +92,7 @@ class ListSharesForFile(APIHandler):
     @gen.coroutine
     def get(self):
         file_path = self.get_query_argument('file_path')
-        yield AsyncRequestHandler.handle_request(self, self.share_api.list_grantees_for_file, 200, file_path)
+        yield RequestHandler.async_handle_request(self, self.share_api.list_grantees_for_file, 200, file_path)
 
 class SharedFolder(APIHandler):
     @property
@@ -102,7 +102,7 @@ class SharedFolder(APIHandler):
     @web.authenticated
     @gen.coroutine
     def get(self):
-        yield AsyncRequestHandler.handle_request(self, self.file_api.list_shared_folder, 200)
+        yield RequestHandler.async_handle_request(self, self.file_api.list_shared_folder, 200)
 
 
 class HomeDirHandler(APIHandler):
@@ -113,7 +113,7 @@ class HomeDirHandler(APIHandler):
     @web.authenticated
     @gen.coroutine
     def get(self):
-        yield AsyncRequestHandler.handle_request(self, self.file_api.get_home_dir, 200)
+        yield RequestHandler.async_handle_request(self, self.file_api.get_home_dir, 200)
 
 class PublicSharesHandler(APIHandler):
     @property
@@ -125,37 +125,37 @@ class PublicSharesHandler(APIHandler):
     def get(self):
         token = self.get_query_argument('token', default=None)
         opaque_id = self.get_query_argument('opaque_id')
-        yield AsyncRequestHandler.handle_request(self, self.public_share_api.get_public_share, 200, opaque_id, token)
+        yield RequestHandler.async_handle_request(self, self.public_share_api.get_public_share, 200, opaque_id, token)
 
     @web.authenticated
     @gen.coroutine
     def post(self):
         request = self.get_json_body()
-        yield AsyncRequestHandler.handle_request(self,
-                                      self.public_share_api.create_public_share,
-                                      201,
-                                      request['endpoint'],
-                                      request['file_path'],
-                                      request['password'],
-                                      request['exp_date'],
-                                      request['permissions'])
+        yield RequestHandler.async_handle_request(self,
+                                                  self.public_share_api.create_public_share,
+                                                  201,
+                                                  request['endpoint'],
+                                                  request['file_path'],
+                                                  request['password'],
+                                                  request['exp_date'],
+                                                  request['permissions'])
 
     @web.authenticated
     @gen.coroutine
     def delete(self):
         opaque_id = self.get_query_argument('opaque_id')
-        yield AsyncRequestHandler.handle_request(self, self.public_share_api.remove_public_share, 204, opaque_id)
+        yield RequestHandler.async_handle_request(self, self.public_share_api.remove_public_share, 204, opaque_id)
 
     @web.authenticated
     @gen.coroutine
     def put(self):
         request = self.get_json_body()
-        yield AsyncRequestHandler.handle_request(self, self.public_share_api.update_public_share,
-                                      204,
-                                      request['opaque_id'],
-                                      request['token'],
-                                      request['field_type'],
-                                      request['field_value'])
+        yield RequestHandler.async_handle_request(self, self.public_share_api.update_public_share,
+                                                  204,
+                                                  request['opaque_id'],
+                                                  request['token'],
+                                                  request['field_type'],
+                                                  request['field_value'])
 
 
 class GetPublicShareByTokenHandler(APIHandler):
@@ -168,7 +168,7 @@ class GetPublicShareByTokenHandler(APIHandler):
     def get(self):
         token = self.get_query_argument('token')
         password = self.get_query_argument('password', default='')
-        yield AsyncRequestHandler.handle_request(self, self.public_share_api.get_public_share_by_token, 200, token, password)
+        yield RequestHandler.async_handle_request(self, self.public_share_api.get_public_share_by_token, 200, token, password)
 
 
 class ListPublicSharesHandler(APIHandler):
@@ -179,7 +179,7 @@ class ListPublicSharesHandler(APIHandler):
     @web.authenticated
     @gen.coroutine
     def get(self):
-        yield AsyncRequestHandler.handle_request(self, self.public_share_api.list_public_shares, 200)
+        yield RequestHandler.async_handle_request(self, self.public_share_api.list_public_shares, 200)
 
 
 class UserInfoHandler(APIHandler):
@@ -192,7 +192,7 @@ class UserInfoHandler(APIHandler):
     def get(self):
         idp = self.get_query_argument('idp')
         opaque_id = self.get_query_argument('opaque_id')
-        yield AsyncRequestHandler.handle_request(self, self.user_api.get_user_info, 200, idp, opaque_id)
+        yield RequestHandler.async_handle_request(self, self.user_api.get_user_info, 200, idp, opaque_id)
 
 class UserInfoClaimHandler(APIHandler):
     @property
@@ -204,7 +204,7 @@ class UserInfoClaimHandler(APIHandler):
     def get(self):
         claim = self.get_query_argument('claim')
         value = self.get_query_argument('value')
-        yield AsyncRequestHandler.handle_request(self, self.user_api.get_user_info_by_claim, 200, claim, value)
+        yield RequestHandler.async_handle_request(self, self.user_api.get_user_info_by_claim, 200, claim, value)
 
 class UserQueryHandler(APIHandler):
     @property
@@ -215,7 +215,7 @@ class UserQueryHandler(APIHandler):
     @gen.coroutine
     def get(self):
         query = self.get_query_argument('query')
-        yield AsyncRequestHandler.handle_request(self, self.user_api.find_users_by_query, 200, query)
+        yield RequestHandler.async_handle_request(self, self.user_api.find_users_by_query, 200, query)
 
 def setup_handlers(web_app, url_path):
     handlers = [
@@ -251,6 +251,19 @@ class RequestHandler(APIHandler):
             RequestHandler.handle_response(self, response, success_code)
 
     @staticmethod
+    async def async_handle_request(self, api_function, success_code, *args):
+        try:
+            loop = get_or_create_eventloop()
+
+            response = await loop.run_in_executor(None, api_function, *args)
+
+        except Exception as err:
+            self.log.error(err)
+            RequestHandler.handle_error(self, err)
+        else:
+            RequestHandler.handle_response(self, response, success_code)
+
+    @staticmethod
     def handle_error(self, err):
         status = RequestHandler.get_response_code(err)
         response = {
@@ -259,52 +272,6 @@ class RequestHandler(APIHandler):
         self.set_header('Content-Type', 'application/json')
         self.set_status(status)
         self.finish(json.dumps(response))
-
-    @staticmethod
-    def handle_response(self, response, success_code):
-        self.set_header('Content-Type', 'application/json')
-        self.set_status(success_code)
-        if response is None:
-            self.finish()
-        else:
-            self.finish(json.dumps(response))
-
-    @staticmethod
-    def get_response_code(err):
-        if isinstance(err, ShareAlreadyExistsError):
-            return 409
-        if isinstance(err, (ShareNotFoundError, LockNotFoundError)):
-            return 404
-        if isinstance(err, (InvalidTypeError, KeyError, FileNotFoundError, ParamError)):
-            return 400
-        if isinstance(err, OCMDisabledError):
-            return 501
-        if isinstance(err, _InactiveRpcError):
-            return 503
-        return 500
-
-class AsyncRequestHandler(APIHandler):
-
-    @staticmethod
-    async def handle_request(self, api_function, success_code, *args):
-        try:
-            loop = asyncio.get_event_loop()
-            response = await loop.run_in_executor(None, api_function, *args)
-
-        except Exception as err:
-            self.log.error(err)
-            AsyncRequestHandler.handle_error(self, err)
-        else:
-            AsyncRequestHandler.handle_response(self, response, success_code)
-
-    @staticmethod
-    def handle_error(self, err):
-        response = {
-            'error_type': err.__class__.__name__,
-            'message': err.message if hasattr(err, 'message') else str(err)
-        }
-        self.set_status(RequestHandler.get_response_code(err))
-        self.finish(json.dumps(str(response)))
 
     @staticmethod
     def handle_response(self, response, success_code):
